feat(timescale_histories): Refactored structure
