+++ pragma-node/src/handlers/onchain/get_entry.rs
@@ -8,10 +8,9 @@ use pragma_entities::EntryError;
 use serde::{Deserialize, Serialize};
 use utoipa::{IntoParams, ToSchema};
 
-use super::get_history::ChunkInterval;
-
-use crate::infra::repositories::onchain_repository::{
-    get_last_updated_timestamp, get_variations, routing, OnchainRoutingArguments,
+use crate::handlers::onchain::get_history::ChunkInterval;
+use crate::infra::repositories::onchain_repository::entry::{
+    routing, OnchainRoutingArguments, get_variations, get_last_updated_timestamp,
 };
 use crate::types::timestamp::TimestampParam;
 use crate::utils::{big_decimal_price_to_hex, PathExtractor};
